requirements.txt
# Basic packages
flask==3.0.2
line-bot-sdk==3.7.0
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0

# Environment variables management
python-dotenv==1.0.1

# Development tools
pytest==8.0.2
black==24.2.0
flake8==7.0.0

# Tunnel for development
pyngrok==7.1.5

# Optional packages

pymongo==4.6.2



schedule==1.2.1
selenium==4.18.1
beautifulsoup4==4.12.3 